(extension) add shared discriminants, seasonLabel, move provider utils to common
- Add isCustomEpisode/isCustomSeason helpers to common/danmaku/utils.ts
- Add isCustom?: true to CustomSeasonInsert in danmaku-converter
- Add seasonSourceKey display helper in common/danmaku/seasonLabel.ts
- Move namespaceKey.ts + resolveSeasonConfig.ts from background/services/providers to common/providers so UI code can import them without crossing the background boundary
- Update all background imports to the new common/providers path
