+++ packages/danmaku-anywhere/src/background/services/matching/strategies/MappingMatchingStrategy.ts
@@ -6,9 +6,9 @@ import type { MatchEpisodeInput, MatchEpisodeResult } from '@/common/anime/dto'
 import { type ILogger, LoggerSymbol } from '@/common/Logger'
 import type { ProviderConfig } from '@/common/options/providerConfig/schema'
 import { ProviderConfigService } from '@/common/options/providerConfig/service'
+import { resolveSeasonConfig } from '@/common/providers/resolveSeasonConfig'
 import { SeasonMap } from '@/common/seasonMap/SeasonMap'
 import { serializeError } from '@/common/utils/serializeError'
-import { resolveSeasonConfig } from '../../providers/resolveSeasonConfig'
 import { EpisodeResolutionService } from '../EpisodeResolutionService'
 import type { IMatchingStrategy } from './IMatchingStrategy'
