+++ packages/danmaku-anywhere/src/common/danmaku/seasonLabel.ts
@@ -1,0 +1,7 @@
+export function seasonSourceKey(season: {
+  manifestId?: string
+  providerConfigId?: string
+  indexedId: string
+}): string {
+  return `${season.manifestId ?? season.providerConfigId ?? 'unknown'}`
+}
