+++ packages/danmaku-anywhere/src/common/providers/namespaceKey.test.ts
